Moved Redis to separate file
