+++ bin/server.js
@@ -8,8 +8,8 @@ const http = require('http')
 const wss = new WebSocket.Server({ noServer: true })
 const setupWSConnection = require('./utils.js').setupWSConnection
 
-const host = process.env.HOST || 'localhost'
-const port = process.env.PORT || 1234
+const host = process.env.YWEBSOCKET_HOST || 'localhost'
+const port = process.env.YWEBSOCKET_PORT || 1234
 
 const server = http.createServer((request, response) => {
   response.writeHead(200, { 'Content-Type': 'text/plain' })
@@ -29,6 +29,8 @@ server.on('upgrade', (request, socket, head) => {
   wss.handleUpgrade(request, socket, head, handleAuth)
 })
 
+console.log(host)
+console.log(port)
 server.listen({ host, port })
 
-console.log(`running at '${host}' on port ${port}`)
+console.log(`collabortate editing server is running at '${host}' on port ${port}`)
